+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Preserve tab group membership when the option to set the focus to the website is used, see [#371](https://github.com/cadeyrn/newtaboverride/issues/371)
 - Improved compatibility with container-aware tab extensions when the option to set the focus to the website is used, see [#217](https://github.com/cadeyrn/newtaboverride/issues/217)
 - Display the selected hex color value for the background color option on the settings page, see [#387](https://github.com/cadeyrn/newtaboverride/issues/387)
+- Added support for configuration via enterprise policies. See the README for details and example, see [#358](https://github.com/cadeyrn/newtaboverride/issues/358)
 
 #### Bugfixes